requirements.txt
Flask~=3.1.3
Werkzeug~=3.1.5
MarkupSafe~=3.0.3
SQLAlchemy~=2.0.46
cryptography~=46.0.5
requests~=2.32.5
flet~=0.9.0
pycryptodome~=3.23.0
setuptools~=82.0.0
beautifulsoup4~=4.14.3
qrcode~=7.4.2
pillow~=12.1.1
Telethon~=1.42.0
python-dotenv~=1.2.1
alembic~=1.18.4
pytest~=9.0.2